Add filtering snapshot, refactor a few things
We now have a "Command" class that knows about dependencies. On top of
this, a (very simple) dependency solver was implemented to ensure that
commands that depend from each other are executed in the correct order.

This is required for example in the snapshots sub-command: A snapshot
may be a filter of another snapshot that may in turn need to be created
first.

The whole framework can (and should) be extended for the rest of the
objects, but currently is only implemented for the snapshots.