Implement generating timestamped snapshots
Snapshots can now have a timestamp component. This is denoted by having
a '%T' placeholder in the name.

Note that there are two syntaxes required in the YAML files: One for
generating snapshots, and one for referencing timestamped snapshot.

Generating timestamps happens by defining a snapshot as usual, but with
a '%T' placeholder. If the placeholder is given, the snapshot object
MUST have a timestamp object denoted below as follows:

  snapshot:
    google-weekly-%T:
      mirror: "google-chrome"
      timestamp:
	# Timestamp spec: If repeat-weekly is present, then the
	# timestamp will be created for the last day matching the given
	# weekday. The time is used to set the time of day when the
	# snapshot "rolls over".
        #
	# Weekdays are denoted in english 3-letter abbreviations. Time
	# is denoted in 24 hour HH:MM format and assumed to be UTC.
        #
        # Example:
        #   - repeat-weekly set to monday, time set to 00:00:
	#      - executed on monday 01:00: a new snapshot is created
	#        labelled with the timestamp of monday (ex.
	#        20151102T0000Z).
	#      - executed on tuesday, 02:00: nothing happens, as there
	#        is already a snapshot matching the "last" repeat date
        repeat-weekly: mon
        time: 00:00

Referencing a snapshot can happen in two ways: Either as a regular
string (if the referenced snapshot is not timestamped), or a dict. If it
is a dict, it MUST have two components: "name", defining the template,
and "timestamp", defining the timestamp configuration (same syntax and
rules as above apply)