Test.yml: new spec for snapshot references.
References to snapshots can be either a single name, or a dict
describing a timestamped snapshot. That dict must have two keys, "name"
and "timestamp". "name" is the timestamp placeholder and must match a
corresponding placeholder name in the snapshots spec.  The "timestamp"
key is a number that references the "age" of timestamp to be referenced:
0 means the most current, 1 the previous one, etc.  Instead of 0, you
may also use the string "current", and instead of 1, you may use the
string "previous".

Example:
   { "name": "google-daily-%T", "timestamp": 0         }