Current state of "snapshot update" implementation.
Note this is currently untested, but (after a lot of tries) the first
half-way sane design.

Instead of building a huge dependency tree of virtual dependencies etc,
it provides "focal point" commands that depend on the commands of the
previous step, thus simplifying the dependency tree immensely. And since
we are modifying the whole snapshot layout during those commands, these
"in-between" commands serve to reload the system state readers knowledge,
thus allowing the upcoming commands to work properly.