Add ignore_existing flag to snapshot and publish cmd functions
This is required as some of those commands may be run when rebuilding the whole
dependency tree of snapshots. Now it may be that *at the planning stage*, some
of the to-be-built publishes or snapshots actually exist, but will be removed
during the rebuilding. Therefore, we need to ignore existing publishes / snaps
during planning.