Fix problem that "previous" references don't reference previous snapshot versions.
The problem is that we're not actually advancing the timestamp back in the
loop; we're just repeating the same calculatin N times.

Note this is a WIP - since my machine is currently kinda broken, I cannot run
the tests myself, thus I'm abusing the PR to let Travis do the job.